+++ frontend/js/state/gens_session.ts
@@ -5,7 +5,18 @@ import {
 import { SideMenu } from "../components/side_menu/side_menu";
 import { COV_Y_RANGE } from "../components/tracks_manager/tracks_manager";
 import { COLORS } from "../constants";
-import { loadAnnotationSelections, loadColorAnnotation, loadCoverageRange, loadExpandedTracks, loadTrackHeights, saveAnnotationSelections, saveColorAnnotation, saveCoverageRange, saveExpandedTracks, saveTrackHeights } from "../util/storage";
+import {
+  loadAnnotationSelections,
+  loadColorAnnotation,
+  loadCoverageRange,
+  loadExpandedTracks,
+  loadTrackHeights,
+  saveAnnotationSelections,
+  saveColorAnnotation,
+  saveCoverageRange,
+  saveExpandedTracks,
+  saveTrackHeights,
+} from "../util/storage";
 import { generateID } from "../util/utils";
 
 /**
@@ -139,7 +150,6 @@ export class GensSession {
   public getTrackExpanded(id: string, defaultValue: boolean): boolean {
     const expanded = this.expandedTracks[id];
     const returnVal = expanded != null ? expanded : defaultValue;
-    console.log(`Return val ${returnVal}`);
     return returnVal;
   }